feat: distribution into modules
