+++ js/util.js
@@ -1,0 +1,12 @@
+const getRandomPositiveInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+const checkStringLength = (string, maxLength) => string.length <= maxLength;
+
+export {checkStringLength, getRandomPositiveInteger, getRandomArrayElement};
